Remove hardcoding of bootloader HRMOR from autocitest
autocitest was doing a comparison between a hardcoded bootloader
HRMOR and the hrmor is read from simics. To make it so we are not
dependent on knowing the original hrmor instead just check that
we are not at a 0x20_0000 offset which indicates we are likely at
HRMOR+2MB, which is where the SBE loads the bootloader.

Change-Id: Ifbc4e8b31c9dc8b06ae07eb4fd9e33540f0d8cbd
RTC: 207396
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/104727
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>