Update default value for HRMOR in ecmd debug framework
Prior to the scratch register being set that tells us the hrmor
we will default to a hard coded value. This commit updates that
hard coded value. Without this change there will be a small window
during the IPL where the tool could use the wrong default value
if the SBE loads us in at the new HRMOR which will be 4GB - 512MB
rather than the previous 4GB - 256MB.

Change-Id: I78aedfd4e1d62bcba41a9e04352a2dc0dad064cf
RTC: 207396
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/104728
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>