Update magic instruction check to compare HRMOR to l3_base
Previously in order to insure that hostboot was running when our
magic instruction handler was triggered we would look at HRMOR
and validate that it matched up with a hard coded value. To get
around needing a hard coded value instead we will look up the
l3_base attribute on the primary proc simics target.

Change-Id: I9056b9af8082fe4b43fd51c752619aa9174a2d71
RTC: 207396
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/104730
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>