Only set l3_base in startup.simics if the SBE scripts have not
In startup.simics we set the l3_base, this tells simics where "0"
is for hostboot. The sbe's startup.simics script also does this for
their standalone model. The SBE is the one who ultimately decides
where Hostboot gets loaded, so if their script has ran and has set
this then we should trust their value instead of overwriting with
what we think is correct.

Change-Id: I9529bc2e46d91821a3c0a1ab3adf199990f04fbe
RTC: 207396
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/104731
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>