Do not explicitly force SBE update for OMI freq diffs
Going to rely on the PLL bucket cross-checks later in the boot
instead of explicitly forcing a SBE update when the OMI frequency
has a mismatch.  Also reordered the logic in call_mss_attr_update
between the forced SBE update and the setting of the PLL bucket
attributes.

At the same time I am removing the old workaround that forces
the OMI frequency to match what the MRW says.

Change-Id: I616cb6dd855725098bb6ca64f92c019aa393ed0d
RTC:258326
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/104789
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>