Fixes 2666 CCS workaround for 4U DDIMMs
CCS writes and PDA fail at 2666 due to the latency
in the Explorer chip. A 0 is written into the DSM register's
WRDATA delay field, which causes CCS data to not be sent.
The fix for this is to increase CWL to an 18.

The workaround for the above assumed that any RCD would have
parity enabled (giving a 4 cycle increased latency). If parity
was enabled, the increased CWL was not needed. However, 4U DDIMM
have an RCD but parity is disabled. This meant that a 0 was still
programmed into the WRDATA delay field and CCS writes/PDA failed.
The fix is to check if the DIMM height is a 4U if we have an RCD.
If it is a 4U, then enable the increased CWL.

Picked from https://rchgit01.rchland.ibm.com/gerrit1/#/c/104804/

Change-Id: I56428dbc1321218420ce06129a60e6eeb9a7a51b
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/104808
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Dev-Ready: STEPHEN GLANCY <sglancy@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Mark Pizzutillo <mark.pizzutillo@ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Louis Stermole <stermole@us.ibm.com>
Reviewed-by: Edgar R Cordero <ecordero@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/104817
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>