Update HWAS::restrictECunits to handle PG containing single & fused cores
There were changes added late in P9 to address problems with
Field Core Override balancing cores on the processors. This
fix broke the use case where we use this function to processes
partial good keyword infomation to deconfigure non-present cores
when we have a mixture of single and fused core pairs represented
in partial good data. This was resulting on the last fused/single core
being dropped from all secondary processors depending if you are in
fused/single mode. This commit adds logic to detect if we are
restricting based on FCO or not and will now only attempt to ensure
the primary processor is allocated cores when doing FCO.

Change-Id: Iecb620ad71d4632b0fa64d8d020a975be46eb281
CQ: SW507893
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/105335
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>