Addressing incompatible FAPI_DBGs on ppe plat
FAPI_DBG is more restrictive on ppe plat than it is on x86. Each of
these files has a couple statments that do not conform with the
requirements for ppe. Most users of the ppe repository do not appear to
compile with FAPI_DBG but it is important for the DFT team, who are now
also users.

We can work around these issues by no passing strings as arguments to
FAPI_DBG and limiting the number of args we pass to a single FAPI_DBG
statement.

Change-Id: I1cfc69f7a4749c1035a85af7231889ab92b9a5bc
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/105370
Reviewed-by: Thi N Tran <thi@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Justin I Smith <justin.i.smith@ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/105381
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>