Fix bug where node offset was being added twice in reserve mem limits
In some recent refactoring we tried to consolidate the logic to lookup
the "base address" hostboot gets loaded at. In some cases this is the
value of the HRMOR register, but in the mirrored memory case this
value need to be adjusted. In this refactor the multi-node case was
overlooked in the code that was verifying the reserved memory limits.
This code was mistakenly applying the node offset twice so all the
checks were failing for multi-node systems.

Change-Id: I0a3003b5d34d8ab4221402f73a6e6e83ea0206a1
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/105420
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>