Fix code in processMrw that calculates OMI_INBAND_BAR_BASE_ADDR_OFFSET
A discrepancy was noticed in the OMI mmio bar settings Hostboot made
in the HW vs what PHYP was thinking they should be. Previously HB kept
incrementing the omi mmio bars for each FAPI_POS. The problem with this
is the FAPI_POS is node-relative and these BAR offsets are relative to
the processor's MMIO space. This commit updates the logic to calculate
bar ranges relative to the processor's mmio space.

Change-Id: I79bd72c5c44e79626b685c0211210a22d1a80cdb
CQ: SW506945
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/105436
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>