Add special procedure for GPIO Devices on the 4U DDIMMs
Testing on HW revealed that the GPIO Expander generic i2c devices
could not be detected with the existing i2cPresence() function like
the PMIC and ADC I2C devices could.  The GPIO devices on the 4U
DDIMMs required a more formal 1-byte read to be put in a good state
after an AC power cycle.  This commit updates the 'detect presence'
path to perform this special procedure on the GPIO devices.  It also
makes some minor changes to the processing of the generic i2c devices
in processMrw.xml.

Change-Id: I13656d8a7aba5aa906ffa28c0c549c07090fbb73
RTC:214692
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/105444
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>