Disable Centaur sensor cache during VPD write request
The Centaur has logic that will constantly do i2c operations out
to the temperature sensors on the dimms.  That bus is shared with
the SPD EEPROM that is used to hold repair information.  There is
a possible contention window if the FSP is writing the SPD while
the Centaur is also using the bus.

To avoid this contention, Hostboot will disable the i2c sensor
cache logic while we are waiting for the FSP to complete a SPD
operation on our behalf.

Change-Id: I1d1fdf59754241534bb4ea41bfe765203ada8830
CQ: SW507279
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/105991
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>