Update CDM_POLICY flag if NO_GARD flag is set in MFG_FLAGS attr
Hostboot looks at the CDM_POLICY attribute to determine if it
should process gard records in PNOR and/or create new gard records
itself. The user should be able to set the MFG_FLAGS attribute to
toggle if they want gard records to apply or not. This commit adds
a check early in hostboot's targeting initialization to make sure
CDM_POLICY matches up with MFG_FLAGS.

Change-Id: Ifa905cbd01e5301c7bb367681adb33821871ce6a
RTC: 248892
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/105801
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>