Fix for seg fault when processing p10.pci.iop.scan.initfile for dd20
This commit fixes processing spy dials that are 1) >64b long, 2) are
in socalled backwards order, and 3) contiguous.

The root cause of the seg fault is in the bit reversal routine that
reverses the bit order of backwards dial bit sequences.  It assumed
a max length of 64b for the bit sequence.  In dd20, however, we
are encountering *contiguous* and *backwards* dials >64b long,
and that has never happened before, not even in p9.

Key_Cronus_Test=XIP_REGRESS_SBE_QME

Change-Id: I8b6513e42fa14a4291b11a0481910ff797adcaf7
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/106022
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Cronus HW CI <cronushw-ci+hostboot@us.ibm.com>
Reviewed-by: Benjamin Gass <bgass@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Claus Michael Olsen <cmolsen@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/106028
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>