Memory encryption support: Zero the first LMB of PAYLOAD memory for PHYP
This commit adds code to fill the first logical memory block (LMB) of
the PAYLOAD with zeroes before starting the payload (when PHYP is in
use).

PHYP depends on the the first LMB of memory being zero when it
starts. We have not had to zero memory explicitly to this point,
because DRAM is already zeroed as part of the memory initialization
steps.

However, when memory encryption is enabled after memory has already
been zeroed, the all-zero blocks of memory will begin to be read
through the decryption engine, and the engine will decrypt a
ciphertext of all zeroes into some unpredictable nonzero plaintext. We
have to re-zero the memory through the encryption engine so that it
actually decrypts into zeroes.

Change-Id: I71a49a1ecceecf8dece97beb64c6f0eb0608084e
RTC: 208820
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/106068
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>