Memory encryption support: Encryption support check, and enable encryption
This commit adds support for checking several sources on whether to
enable memory encryption, and also enables encryption by default via
the HWP attribute.

Encryption should not be enabled if export controls restrict the use
of cryptography. If there are no export controls, then encryption
should be enabled (1) if secureboot is on, or (2) if an attribute is
set to enable it.

Change-Id: If55e6ade6a10ad315e43085dfaef648c425795ec
RTC: 208820
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/106097
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>