Memory encryption support: Zero HB Reserved Memory region
This commit adds support for zeroing the HB Reserved Memory region
after exiting cache-contained mode.

This was unnecessary before because all of DRAM was zeroed as part of
memory initialization. However, after memory encryption is enabled,
the previously-zeroed memory gets "decrypted" as garbage, so we have
to re-zero the memory through the encryption engine so that it
decrypts to zeroes.

Change-Id: I3c50260e00dfce6c2daad1d06d0eeccbd4bf3127
RTC: 208820
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/106148
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Christopher J Engel <cjengel@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>