Memory encryption support: Zero the first LMB of PAYLOAD memory for OPAL
When loading PHYP with memory encryption enabled, Hostboot is required
to zero the first logical memory block (LMB) of PAYLOAD memory. This
commit causes OPAL to be treated the same as PHYP in this regard.
While this was not a direct requirement for memory encryption support,
uniform treatment of the two hypervisors simplifies the code and
provides a stronger guarantee about the state of memory to OPAL.

Change-Id: I7fa4df7182bb15800c59ff610c5d2fe3e703f692
RTC: 208820
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/106154
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Christopher J Engel <cjengel@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>