Memory encryption support: Random number generation algorithm
This commit introduces the basic algorithm for creating 128-bit
cryptographic keys from a hardware-based 32-bit random number
generator via the DARN ("Deliver a Random Number") instruction.

This commit does not enable DARN usage yet, because it depends on a
base-address register in the RNG hardware to be initialized, and the
HWP to do this is not available yet.

Change-Id: If43faa125198ab5cacd973c147196a76ae41c7eb
RTC: 208820
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/106163
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Christopher J Engel <cjengel@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>