Fix for defect SW509480
- Fix for SW509480.  During MPIPL the fabric configuration is
cleared, causing the Slave TODs to no longer receive Sync pulses
from the MDMT, and they report a "special status" error as a
result.  This special status error prevents the Slave TOD FSMs
from indicating the Running state during the later TOD init sequence.
The fix is to move the fabric configuration clearing past the
point where the Step checkers of the Slave TODs are disabled.
- Fixed a related issue where if the special status error is
already set in the Slave TODs before the TOD setup sequence
is invoked during MPIPL, then the TOD init sequence would
fail in the Slave TODs because the special status error
was not cleared.
- Patch Set 2: Removed redundant clearing of Fabric DL TOD inits.

CQ: SW509480
Change-Id: I14d0af3f3d504a6188610dac7e3eb8c3d42f0918
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/106277
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Reviewed-by: Manish K Chowdhary <manichow@in.ibm.com>
Reviewed-by: Rajees R P P <rajerpp1@in.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: NAREN A DEVAIAH <naren.devaiah@in.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/106279
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>