Fix invalid FSP_BASE_ADDR on node1
There was a big in the topology id to index conversion logic
in the MRW parser script that resulted in the second node seeing
an index of "8" instead of "4" for ATTR_FSP_BASE_ADDR.  This led
to CRESP errors when PHYP tried to access the FSP via the expected
BAR.

The conversion function is fixed, however it is also no longer
being called.  Instead we will set the value of the attribute
inside Hostboot code in the same way we handle all of the other
BARs since they are remapped when the effective topology id
changes.

Change-Id: I2abee2a29d44dc19e5446a24d68df263837b4ecd
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/106331
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>