Remove ekb scratch attributes
We do not want to instantiate the fapi scratch attributes because
they add 100s of KB of to our potential memory footprint.  For
example, on a 2-proc, 16-dimm configuration the amount of space
needed for the HEAP attributes dropped from 332KB to 64KB with
this change.

Keeping a few on the system target for testcases.

Change-Id: Iabe6e78f199193ec6d1b650a88a7305abab150e8
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/106835
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>