XSCOM Driver in HBBL: xscomPerformOp Interface
This commit adds a xscomPerformOp interface and implementation to HBBL.
xscomPerformOp is capable of performing XSCOM (SCOM local to the chip)
read and write OPs. Note that all operations are double-word in size,
but 32-bit registers can also be read via xscomPerformOp. Some assembly
changes were required in handleMMIO to be able to write larger pieces of
data.

Change-Id: Ic76140560033db4cd48f8b8ac4ddaf41c9d391d1
RTC: 216130
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/107729
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>