Memory encryption support: Prevent cryptographic keys from being traced
This commit adds a new interface that code can use to prevent the
values of SCOM reads and writes from being traced out in debug traces
in the SCOM handler routines. (Right now only XSCOM operations support
it, but we can add others in the future.)

This is used by the memory encryption setup code so that if we somehow
release a Hostboot build with SCOM value traces enabled, the values of
the cryptographic keys will still not be leaked via the traces.

Change-Id: Iafd87d8b31a448fdd2e31fc116f9260b75be5385
RTC: 208820
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/106931
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>