Add the no-op APIs to update/validate the VPD ECC data
This will add the infrastructure to update and validate the VPD ECC data.
Currently these are no-ops, but can easily be updated, via a patch, with
the actual code to implement these APIs, with very minimal impact on the
current code.

The module/reason IDs are added in this commit so hat they can be picked
up in the patch that provides the implementation.

Change-Id: I284b59da5d7e62369b6cade97c21741b5f9f951b
RTC: 250100
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/107173
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>