Disable memory encryption
Currently, memory encryption is activated in istep 14. This was found
to cause a checkstop intermittently. The prevailing theory is that the
memory hardware has a "race condition" of sorts, which exposes us to a
SCOM parity error when memory encryption is enabled while the memory
hardware is concurrently servicing a SCOM operation.

The experimental solution is to move p10_init_mem_encryption to istep
12, when SCOM operations are still serviced by I2C.

This commit temporarily disables memory encryption until we can move
the HWP call to the correct istep.

Change-Id: Ib5e474147bdd9a0b614e106eae8ceef4de89eb16
RTC: 208820
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/107412
Reviewed-by: Adam S Hale <adam.samuel.hale@ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christopher J Engel <cjengel@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>