When MMIO i/o errors to the OCMBs are found switch to using SBE scoms
Previously we would switch over to doing i2c scoms from the host
in when we found i/o ops but there is a security option that can
disable this path. In order to have a safe and consistent backup
when MMIO fails we will always fall back to asking the SBE to
perform the i2c scoms to the OCMB with a PSU op instead of doing
the i2c scom from the host.

Change-Id: I3dde410c314e84d97a4390bb95052618eb7a4b0f
RTC: 195684
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/107544
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>