Fixed race condition in RAMing related to HW533775
Fixed a race condition in the RAMing code related to the DD1
workaround for HW533775 (Worthwhile Toad, recovery invocation).
This workaround requires reading the VTID_V bits from the
THREAD_INFO register.
When a RAMing operation ends, it may deactivate a thread,
and the VTID_V bit can take a long time to deassert, and may
remain asserted until the next RAMing operation begins.
If the next RAMing operation reads the VTID_V bit and it
still sees VTID_V asserted from the previous RAMing operation,
the code may mistakenly determine that the workaround for
HW533775 is required, and it will invoke the recovery
mechanism, instead of skipping recovery.  Meanwhile the
HW can deassert the VTID_V bit after the HWP reads it,
and the result will be a situation where there are no VT
mappings present when recovery is invoked by the HWP.
There is a HW known issue with invoking recovery when no
VT mappings are present (commit 100185), and so the second
RAMing operation fails.
The fix is to always wait for the THREAD_ACTION_IN_PROGRESS
bit 23 in the THREAD_INFO register to be deasserted before
considering the contents of the THREAD_INFO register to be valid.

Change-Id: Ie4a977bd4973ec526bef1fd1cf8f6565e7248315
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/107633
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Thi N Tran <thi@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/107638
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>