Use TARGETING::PRESENT for TARGET_STATE_PRESENT for chip targets
There is translation required to go from the binary functional
representation of fapi2::TargetState to the trinary state that
we have inside of Hostboot.  For chiplet/subunits, the ekb
code expects TARGET_STATE_PRESENT to represent the architectural
limits of the chip, meaning that we need to include parts marked
as bad in PG.  Since those are marked non-present in TARGETING
we will use the ALL (aka. blueprint) value.  However, for top-level
targets (chips, dimms), PRESENT should represent what is physically
installed.

Change-Id: Id9e7e8b5024c09a919f3424380d00c3ad0ad7d69
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/107828
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>