Add toggle of pmic efuse enable to fix n-mode degrade
Through some testing, it appears the efuse latch circuit is
getting stuck in an off state, which remains stuck when
turning the efuse on. The reason for this seems to vary,
and it seems to occur on Denali systems during inital
power on. It results in no 12V being detected at the PMIC.

It appears as though toggling the efuse on/off/on gets
the reset line of the latch to assert out of the stuck
state, resulting in consistently good runs on our
test benches without the N-Mode failure.

(cherry picked from commit 6a91353824847109a32c798802089781b16fe747)

Change-Id: I495f42636b9ce8361ebc6381e866f9515a715f12
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/108951
Reviewed-by: Louis Stermole <stermole@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: STEPHEN GLANCY <sglancy@us.ibm.com>
Dev-Ready: STEPHEN GLANCY <sglancy@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Edgar R Cordero <ecordero@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/108964
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>