EECACHE synchronization for VPD
Hostboot now owns VPD/EEPROM content (processor module seeproms
and DDIMM SPD).  Hostboot as owner of the EEPROMs can write
directly to the SEEPROM hardware (as well as in memory eecache
content).

The FSP vpd cache needs to be notified when VPD content is
modified via sendMboxWriteMsg's.  The EECACHE partition
also needs to be flagged as needing refreshed at the next
Hostboot IPL.

The existing FSP messaging infrastructure will be used
(new VPD_WRITE_CACHE type) to notifiy the FSP that the HUID
target VPD has been updated.

A new attribute will be added that will be set on targets
to identify that the VPD data (eecache in memory and SEEPROM
hardware) has been modified.

This new attribute will be checked during Hostboot IPL
to properly trigger the updating of the EECACHE partition.

Change-Id: I64a4e4cd6fb12ef39a495ed54d0d78e6cdcf67e9
RTC:251157
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/108466
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>