SBE Update: Don't try to read via SPI if ChipOp already failed
This commit removes the code that tried to read the SBE Seeprom's
Version Information Struct via SPI if it first failed to read it
via a SBE ChipOp.  It was decided that for p10 it wasn't worth
risking a SPI bus collision if the SBE ChipOp already failed.
There was some additional code cleanup, as well.

Change-Id: I2aed5cd283fe0cbb4fec2d2e6f6badb1b669d75d
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/108512
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>