New SBE PSU "unsupported command" error handling (commit 1/2)
This commit adds automatic handling for "unsupported command" errors
for SBE PSU operations.

Previously the caller of performPsuChipOp needed to check the return
codes to tell whether the SBE's response was "command not supported."
The new interface makes "command not supported" an unrecoverable error
by default, and gives callers two options: (1) they can tone down the
severity of the error, by providing a severity themselves, and (2)
they can provide a boolean pointer, which will be set to true or false
depending on whether the command was supported, if they need more
fine-grained handling of the situation.

RTC: 173219
Change-Id: Ib68571a1f796acd6a38675b75f443ee70580600f
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/108553
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>